dcrwallet rpc corrections
