+++ web/static/app/src/controllers/connection-info_controller.js
@@ -14,17 +14,20 @@ export default class extends Controller {
   }
 
   connect () {
-    let _this = this
+    const blocksRescanProgress = localStorage.getItem('blocks-rescan-progress')
+    if (blocksRescanProgress) {
+      this.blockScanProgressTarget.textContent = blocksRescanProgress
+    }
 
-    ws.registerEvtHandler('updateConnInfo', function (data) {
-      _this.peersConnectedTarget.textContent = data.peersConnected
-      _this.totalBalanceTarget.textContent = data.totalBalance
-      _this.latestBlockTarget.textContent = data.latestBlock
-      _this.networkTypeTarget.textContent = data.networkType
+    ws.registerEvtHandler('updateConnInfo', data => {
+      this.peersConnectedTarget.textContent = data.peersConnected
+      this.totalBalanceTarget.textContent = data.totalBalance
+      this.latestBlockTarget.textContent = data.latestBlock
+      this.networkTypeTarget.textContent = data.networkType
     })
 
-    ws.registerEvtHandler('updateBalance', function (data) {
-      _this.totalBalanceTarget.textContent = data.total
+    ws.registerEvtHandler('updateBalance', data => {
+      this.totalBalanceTarget.textContent = data.total
     })
 
     ws.registerEvtHandler('updateSyncProgress', syncInfo => {
@@ -32,12 +35,15 @@ export default class extends Controller {
       // or if block headers rescan has not started or has completed (i.e. syncInfo.rescanProgress <= 0 || syncInfo.rescanProgress >= 100)
       if (!syncInfo.done || syncInfo.rescanProgress <= 0 || syncInfo.rescanProgress >= 100) {
         hide(this.blockScanProgressTarget)
+        localStorage.removeItem('blocks-rescan-progress')
         return
       }
 
       this.blockScanProgressTarget.textContent = `Rescanning block headers ${syncInfo.rescanProgress}%. `
       this.blockScanProgressTarget.textContent += `Scanning ${syncInfo.currentRescanHeight} of ${syncInfo.totalHeadersToFetch} block headers.`
       show(this.blockScanProgressTarget)
+
+      localStorage.setItem('blocks-rescan-progress', this.blockScanProgressTarget.textContent)
     })
   }
 }
