+++ web/views/layout.html
@@ -31,7 +31,7 @@
                         | Latest Block: <b data-target="connection-info.latestBlock">{{ .LatestBlock }}</b>
                     </p>
                     <!-- block rescan progress display, ideally entire blockchain sync progress should persist on all pages like this -->
-                    <p class="mb-0 d-none" data-target="connection-info.blockScanProgress"></p>
+                    <p id="blocks-rescan-progress" class="mb-0 d-none" data-target="connection-info.blockScanProgress"></p>
                 </div>
             </div>
         </div>
